Update 01.ps1
In a given text (let it be set as a variable)  find text "Hello World". If it exists in the text, then replace all "Hello World"-s to "Happy New Year" and calculate count of a word "New" in the result text. If there is no "Hello World" in the text then save it as a text file and  add to the title of the file number of spaces in the text.